+++ logprep/processor/key_checker/rule.py
@@ -62,3 +62,6 @@ class KeyCheckerRule(FieldManagerRule):
             converter=set,
         )
         """List of fields to check for."""
+        target_field: str = field(validator=validators.instance_of(str))
+        """The field where to write the processed values to. """
+        mapping: dict = field(default="", init=False, repr=False, eq=False)