+++ tests/unit/processor/field_manager/test_field_manager.py
@@ -282,6 +282,126 @@ test_cases = [  # testcase, rule, event, expected
             },
         },
     ),
+    (
+        "copies multiple fields to multiple target fields",
+        {
+            "filter": "field",
+            "field_manager": {
+                "mapping": {"field.one": "one", "field.two": "two", "field.three": "three"},
+            },
+        },
+        {"field": {"one": 1, "two": 2, "three": 3}},
+        {"field": {"one": 1, "two": 2, "three": 3}, "one": 1, "two": 2, "three": 3},
+    ),
+    (
+        "copies multiple fields to multiple target fields, while overwriting existing fields",
+        {
+            "filter": "field",
+            "field_manager": {
+                "mapping": {"field.one": "one", "field.two": "two", "field.three": "three"},
+                "overwrite_target": True,
+            },
+        },
+        {"field": {"one": 1, "two": 2, "three": 3}, "three": "exists already"},
+        {"field": {"one": 1, "two": 2, "three": 3}, "one": 1, "two": 2, "three": 3},
+    ),
+    (
+        "copies multiple fields to multiple target fields, while one list will be extended",
+        {
+            "filter": "field",
+            "field_manager": {
+                "mapping": {"field.one": "one", "field.two": "two", "field.three": "three"},
+                "extend_target_list": True,
+            },
+        },
+        {"field": {"one": 1, "two": 2, "three": 3}, "three": ["exists already"]},
+        {
+            "field": {"one": 1, "two": 2, "three": 3},
+            "one": 1,
+            "two": 2,
+            "three": ["exists already", 3],
+        },
+    ),
+    (
+        "copies multiple fields to multiple target fields, while one list will be extended with existing list",
+        {
+            "filter": "field",
+            "field_manager": {
+                "mapping": {"field.one": "one", "field.two": "two", "field.three": "three"},
+                "extend_target_list": True,
+            },
+        },
+        {"field": {"one": 1, "two": 2, "three": [3, 3]}, "three": ["exists already"]},
+        {
+            "field": {"one": 1, "two": 2, "three": [3, 3]},
+            "one": 1,
+            "two": 2,
+            "three": ["exists already", 3, 3],
+        },
+    ),
+    (
+        "copies multiple fields to multiple target fields, while one target list will be overwritten with existing list",
+        {
+            "filter": "field",
+            "field_manager": {
+                "mapping": {"field.one": "one", "field.two": "two", "field.three": "three"},
+                "overwrite_target": True,
+            },
+        },
+        {"field": {"one": 1, "two": 2, "three": [3, 3]}, "three": ["exists already"]},
+        {"field": {"one": 1, "two": 2, "three": [3, 3]}, "one": 1, "two": 2, "three": [3, 3]},
+    ),
+    (
+        "copies multiple fields to multiple target fields, while one source field is missing",
+        {
+            "filter": "field",
+            "field_manager": {
+                "mapping": {"field.one": "one", "field.two": "two", "field.three": "three"},
+            },
+        },
+        {
+            "field": {"one": 1, "three": 3},
+        },
+        {
+            "field": {"one": 1, "three": 3},
+            "one": 1,
+            "three": 3,
+            "tags": ["_field_manager_missing_field_warning"],
+        },
+    ),
+    (
+        "moves multiple fields to multiple target fields",
+        {
+            "filter": "field",
+            "field_manager": {
+                "mapping": {"field.one": "one", "field.two": "two", "field.three": "three"},
+                "delete_source_fields": True,
+            },
+        },
+        {"field": {"one": 1, "two": 2, "three": 3}},
+        {"one": 1, "two": 2, "three": 3},
+    ),
+    (
+        "Combine fields to list and copy fields at the same time",
+        {
+            "filter": "field",
+            "field_manager": {
+                "source_fields": ["source.one", "source.two"],
+                "target_field": "merged",
+                "mapping": {"field.one": "one", "field.two": "two", "field.three": "three"},
+                "extend_target_list": True,
+            },
+        },
+        {"field": {"one": 1, "two": 2, "three": 3}, "source": {"one": ["a"], "two": ["b"]}},
+        {
+            "field": {"one": 1, "two": 2, "three": 3},
+            "source": {"one": ["a"], "two": ["b"]},
+            "one": 1,
+            "two": 2,
+            "three": 3,
+            "merged": ["a", "b"],
+        },
+    ),
 ]
 
 failure_test_cases = [
@@ -295,7 +415,8 @@ failure_test_cases = [
             },
         },
         {"message": "This is a message"},
-        {"message": "This is a message", "tags": ["_field_manager_failure"]},
+        {"message": "This is a message", "tags": ["_field_manager_missing_field_warning"]},
+        ".*ProcessingWarning.*",
     ),
     (
         "single source field not found and preexisting tags",
@@ -307,7 +428,11 @@ failure_test_cases = [
             },
         },
         {"message": "This is a message", "tags": ["preexisting"]},
-        {"message": "This is a message", "tags": ["_field_manager_failure", "preexisting"]},
+        {
+            "message": "This is a message",
+            "tags": ["_field_manager_missing_field_warning", "preexisting"],
+        },
+        ".*ProcessingWarning.*",
     ),
     (
         "single source field not found and preexisting tags with deduplication",
@@ -318,10 +443,35 @@ failure_test_cases = [
                 "target_field": "new_field",
             },
         },
-        {"message": "This is a message", "tags": ["_field_manager_failure", "preexisting"]},
-        {"message": "This is a message", "tags": ["_field_manager_failure", "preexisting"]},
+        {
+            "message": "This is a message",
+            "tags": ["_field_manager_missing_field_warning", "preexisting"],
+        },
+        {
+            "message": "This is a message",
+            "tags": ["_field_manager_missing_field_warning", "preexisting"],
+        },
+        ".*ProcessingWarning.*",
     ),
-]  # testcase, rule, event, expected
+    (
+        "copies multiple fields to multiple target fields, while one target exists already",
+        {
+            "filter": "field",
+            "field_manager": {
+                "mapping": {"field.one": "one", "field.two": "two", "field.three": "three"},
+            },
+        },
+        {"field": {"one": 1, "two": 2, "three": 3}, "three": "exists"},
+        {
+            "field": {"one": 1, "two": 2, "three": 3},
+            "one": 1,
+            "two": 2,
+            "three": "exists",
+            "tags": ["_field_manager_failure"],
+        },
+        ".*FieldExistsWarning.*",
+    ),
+]  # testcase, rule, event, expected, error
 
 
 class TestFieldManager(BaseProcessorTestCase):
@@ -337,12 +487,12 @@ class TestFieldManager(BaseProcessorTestCase):
         self.object.process(event)
         assert event == expected
 
-    @pytest.mark.parametrize("testcase, rule, event, expected", failure_test_cases)
-    def test_testcases_failure_handling(self, testcase, rule, event, expected, caplog):
+    @pytest.mark.parametrize("testcase, rule, event, expected, error", failure_test_cases)
+    def test_testcases_failure_handling(self, testcase, rule, event, expected, error, caplog):
         self._load_specific_rule(rule)
         with caplog.at_level(logging.WARNING):
             self.object.process(event)
-        assert re.match(".*ProcessingWarning.*", caplog.text)
+        assert re.match(error, caplog.text)
         assert event == expected, testcase
 
     def test_process_raises_duplication_error_if_target_field_exists_and_should_not_be_overwritten(