+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 notation.
  
 ### Improvements
+
+* Extend the `FieldManager` such that it can move/copy multiple source fields into multiple targets
+inside one rule.
+
 ### Bugfix
 
 * Fix error handling of missing source fields in grokker